+++ app/src/main/java/androidx/preference/CertPickerPreference.java
@@ -8,7 +8,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.ubergeek42.WeechatAndroid.service.SSLHandler;
-import com.ubergeek42.WeechatAndroid.utils.SecurityUtils;
+import com.ubergeek42.WeechatAndroid.utils.AndroidKeyStoreUtils;
 
 public class CertPickerPreference extends PasswordedFilePickerPreference implements DialogFragmentGetter {
     public CertPickerPreference(Context context, AttributeSet attrs) {
@@ -22,7 +22,7 @@ public class CertPickerPreference extends PasswordedFilePickerPreference impleme
         if (bytes == null) {
             return "Certificate forgotten";
         } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            return SecurityUtils.areAllInsideSecurityHardware(SSLHandler.KEYSTORE_ALIAS_PREFIX) ?
+            return AndroidKeyStoreUtils.areAllInsideSecurityHardware(SSLHandler.KEYSTORE_ALIAS_PREFIX) ?
                     "Certificate imported into security hardware" :
                     "Certificate imported into a software key store";
         } else {
